completed task 11 - changed minor query detail

+++ models/articles.models.js
@@ -33,7 +33,7 @@ exports.fetchAllArticles = (topic, sort_by = "created_at", order = "desc") => {
   }
   let text = "";
   const params = [];
-  text += `SELECT articles.author, articles.title, articles.article_id, articles.topic, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.comment_id) AS comment_count FROM articles LEFT JOIN comments ON articles.article_id = comments.article_id `;
+  text += `SELECT articles.author, articles.title, articles.article_id, articles.topic, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.comment_id)::INTEGER AS comment_count FROM articles LEFT JOIN comments ON articles.article_id = comments.article_id `;
 
   if (topic) {
     text += `WHERE topic = $1 `;
